Others/bot_behavior/requirements.txt
discord
python-dotenv
requests
redis
